Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/intersection-observer/same-document-with-document-root-expected.txt


PASS IntersectionObserver in a single document using document as root.
PASS First rAF.
FAIL document.scrollingElement.scrollTop = 300 assert_equals: entries.length expected 2 but got 1
